usnic: Always prefix on recv (secretly, shhh).
Trying to get around the issue in the shared PD case where the
truncation path attempts to read the header, but the address it finds is
the iova address that is not usable. In this case, the header is
allocated by the application so it is not problematic. The completion
lengths get touched up, so fix that too. The application doesn't know
we're secrelty using prefix, so remove the side of the 42-byte header.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>